requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
pymongo
pinecone-client
langchain
langgraph
langsmith
openai
pydantic
pydantic-settings
pypdf
pillow
pytesseract
pymupdf
langchain_anthropic
langchain_openai
langchain-community
pinecone-client
langchain-text-splitters 
sentence-transformers
requests
python-jose
passlib[bcrypt]
streamlit requests
pydantic[email]